requirements.txt
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
Flask==3.0.0
influxdb-client==1.38.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
python-dateutil==2.8.2
reactivex==4.0.4
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
